+++ src/lib.rs
@@ -4,6 +4,11 @@
 #![deny(missing_docs)]
 #![cfg_attr(docsrs, feature(doc_auto_cfg))]
 
+pub use pep440_rs;
+
+/// Utilities for working with OME-Zarr version strings.
+pub(crate) mod version;
+
 /// Version `0.4` (OME-NGFF) metadata.
 ///
 /// <https://ngff.openmicroscopy.org/0.4/>.
@@ -17,6 +22,14 @@ pub mod v0_5;
 /// Version `0.6` metadata.
 pub mod v0_6;
 
+/// Future version of OME-Zarr metadata.
+///
+/// May not be up to date with the latest development specification.
+/// May implement not-yet-stabilised RFCs.
+/// Breaking API changes may be made at any time.
+#[cfg(feature = "next")]
+pub mod next;
+
 mod errors;
 pub use errors::{Error, Result};
 
@@ -24,3 +37,6 @@ mod ndim;
 pub use ndim::{MaybeNDim, NDim};
 
 pub use validatrix::{Valid, Validate};
+
+mod any;
+pub use any::AnyOmeFields;