+++ src/v0_4.rs
@@ -18,7 +18,20 @@ use serde::{Deserialize, Serialize};
 use validatrix::{Accumulator, Validate};
 pub use well::*;
 
+/// Alias for [OmeNgffGroupAttributes] for consistency with later versions
+/// where the OME-Zarr fields are namespaced.
+pub type OmeFields = OmeNgffGroupAttributes;
+
+/// Alias for [OmeNgffGroupAttributes] for consistency with later versions
+/// (OME-NGFF was renamed to OME-Zarr coinciding with v0.5).
+pub type OmeZarrGroupAttributes = OmeNgffGroupAttributes;
+
+crate::constrained_version!(ConstrainedVersion, "==0.4", "0.4");
+
 /// OME-NGFF top-level group attributes.
+///
+/// Has aliases [OmeFields] and [OmeZarrGroupAttributes] for consistency with later versions.
+/// In later versions, the OME fields are namespaced and so those two types refer to different things.
 #[derive(Serialize, Deserialize, Debug, Clone, Default)]
 pub struct OmeNgffGroupAttributes {
     /// Transitional `bioformats2raw.layout` metadata.
@@ -46,6 +59,32 @@ pub struct OmeNgffGroupAttributes {
     /// Transitional OMERO metadata.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub omero: Option<Omero>,
+}
+
+impl OmeNgffGroupAttributes {
+    /// Get the first available version string for the OME-NGFF metadata.
+    ///
+    /// Checks `multiscales`, `image-label`, `plate`, and `well` fields in that order; falls back to `"0.4"`.
+    pub fn version(&self) -> String {
+        if let Some(v) = self
+            .multiscales
+            .as_ref()
+            .and_then(|ms| ms.first())
+            .map(|ms| ms.version.to_string())
+        {
+            return v;
+        }
+        if let Some(v) = self.image_label.as_ref().map(|ms| ms.version.to_string()) {
+            return v;
+        }
+        if let Some(v) = self.plate.as_ref().map(|ms| ms.version.to_string()) {
+            return v;
+        }
+        if let Some(v) = self.well.as_ref().map(|ms| ms.version.to_string()) {
+            return v;
+        }
+        "0.4".into()
+    }
 }
 
 impl Validate for OmeNgffGroupAttributes {